Refactor authentication handling by moving AuthCallback to a new component and implementing user state management in App component.
